+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaUserQuotas.java
@@ -34,15 +34,15 @@ public class KafkaUserQuotas implements UnknownPropertyPreserving, Serializable 
     private Integer producerByteRate;
     private Integer consumerByteRate;
     private Integer requestPercentage;
-    private Integer controllerMutationRate;
+    private Double controllerMutationRate;
 
-    @Description(" The quota on rate at which mutations are accepted for the create topics request, the create partitions request and the delete topics request. The rate is accumulated by the number of partitions created or deleted.")
+    @Description("A quota on the rate at which mutations are accepted for the create topics request, the create partitions request and the delete topics request. The rate is accumulated by the number of partitions created or deleted.")
     @Minimum(0)
-    public Integer getControllerMutationRate() {
+    public Double getControllerMutationRate() {
         return controllerMutationRate;
     }
 
-    public void setControllerMutationRate(Integer controllerMutationRate) {
+    public void setControllerMutationRate(Double controllerMutationRate) {
         this.controllerMutationRate = controllerMutationRate;
     }