+++ user-operator/src/test/java/io/strimzi/operator/user/operator/KafkaUserQuotasIT.java
@@ -76,7 +76,7 @@ public class KafkaUserQuotasIT {
         defaultQuotas = new KafkaUserQuotas();
         defaultQuotas.setConsumerByteRate(1000);
         defaultQuotas.setProducerByteRate(2000);
-        defaultQuotas.setControllerMutationRate(10);
+        defaultQuotas.setControllerMutationRate(10d);
     }
 
     @Test
@@ -131,7 +131,7 @@ public class KafkaUserQuotasIT {
         KafkaUserQuotas newQuotas = new KafkaUserQuotas();
         newQuotas.setConsumerByteRate(1000);
         newQuotas.setProducerByteRate(2000);
-        newQuotas.setControllerMutationRate(10);
+        newQuotas.setControllerMutationRate(10d);
         kuq.createOrUpdate(Reconciliation.DUMMY_RECONCILIATION, username, newQuotas);
         assertThat(isPathExist("/config/users/" + encodeUsername(username)), is(true));
         testDescribeUserQuotas(username, newQuotas);
@@ -218,9 +218,9 @@ public class KafkaUserQuotasIT {
     public void testUpdateControllerMutationRate() throws Exception {
         String username = "changeControllerMutationRate";
         kuq.createOrUpdate(Reconciliation.DUMMY_RECONCILIATION, username, defaultQuotas);
-        defaultQuotas.setControllerMutationRate(20);
+        defaultQuotas.setControllerMutationRate(20d);
         kuq.createOrUpdate(Reconciliation.DUMMY_RECONCILIATION, username, defaultQuotas);
-        assertThat(kuq.describeUserQuotas(Reconciliation.DUMMY_RECONCILIATION, username).getControllerMutationRate(), is(20));
+        assertThat(kuq.describeUserQuotas(Reconciliation.DUMMY_RECONCILIATION, username).getControllerMutationRate(), is(20d));
     }
 
     @Test
@@ -229,7 +229,7 @@ public class KafkaUserQuotasIT {
         quotas.setConsumerByteRate(2000);
         quotas.setProducerByteRate(4000);
         quotas.setRequestPercentage(40);
-        quotas.setControllerMutationRate(10);
+        quotas.setControllerMutationRate(10d);
         Set<ClientQuotaAlteration.Op> ops = kuq.toClientQuotaAlterationOps(quotas);
         assertThat(ops, hasSize(4));
         assertThat(ops.contains(new ClientQuotaAlteration.Op("consumer_byte_rate", 2000d)), is(true));
@@ -262,7 +262,7 @@ public class KafkaUserQuotasIT {
         assertThat(quotas.getConsumerByteRate(), is(2000));
         assertThat(quotas.getProducerByteRate(), is(4000));
         assertThat(quotas.getRequestPercentage(), is(40));
-        assertThat(quotas.getControllerMutationRate(), is(10));
+        assertThat(quotas.getControllerMutationRate(), is(10d));
 
         map.remove("consumer_byte_rate");
         map.remove("producer_byte_rate");
@@ -291,7 +291,7 @@ public class KafkaUserQuotasIT {
         quotas.setConsumerByteRate(2_000_000);
         quotas.setProducerByteRate(1_000_000);
         quotas.setRequestPercentage(50);
-        quotas.setControllerMutationRate(10);
+        quotas.setControllerMutationRate(10d);
 
         assertThat(kuq.exists(Reconciliation.DUMMY_RECONCILIATION, username), is(false));
 
@@ -320,7 +320,7 @@ public class KafkaUserQuotasIT {
         initialQuotas.setConsumerByteRate(2_000_000);
         initialQuotas.setProducerByteRate(1_000_000);
         initialQuotas.setRequestPercentage(50);
-        initialQuotas.setControllerMutationRate(10);
+        initialQuotas.setControllerMutationRate(10d);
 
         kuq.createOrUpdate(Reconciliation.DUMMY_RECONCILIATION, username, initialQuotas);
         assertThat(kuq.exists(Reconciliation.DUMMY_RECONCILIATION, username), is(true));
@@ -330,7 +330,7 @@ public class KafkaUserQuotasIT {
         updatedQuotas.setConsumerByteRate(4_000_000);
         updatedQuotas.setProducerByteRate(3_000_000);
         updatedQuotas.setRequestPercentage(75);
-        updatedQuotas.setControllerMutationRate(10);
+        updatedQuotas.setControllerMutationRate(10d);
 
         Checkpoint async = testContext.checkpoint();
         kuq.reconcile(Reconciliation.DUMMY_RECONCILIATION, username, updatedQuotas)
@@ -357,7 +357,7 @@ public class KafkaUserQuotasIT {
         initialQuotas.setConsumerByteRate(2_000_000);
         initialQuotas.setProducerByteRate(1_000_000);
         initialQuotas.setRequestPercentage(50);
-        initialQuotas.setControllerMutationRate(10);
+        initialQuotas.setControllerMutationRate(10d);
 
         kuq.createOrUpdate(Reconciliation.DUMMY_RECONCILIATION, username, initialQuotas);
         assertThat(kuq.exists(Reconciliation.DUMMY_RECONCILIATION, username), is(true));
@@ -366,7 +366,7 @@ public class KafkaUserQuotasIT {
         KafkaUserQuotas updatedQuotas = new KafkaUserQuotas();
         updatedQuotas.setConsumerByteRate(4_000_000);
         updatedQuotas.setProducerByteRate(3_000_000);
-        updatedQuotas.setControllerMutationRate(20);
+        updatedQuotas.setControllerMutationRate(20d);
 
         Checkpoint async = testContext.checkpoint();
         kuq.reconcile(Reconciliation.DUMMY_RECONCILIATION, username, updatedQuotas)
@@ -394,7 +394,7 @@ public class KafkaUserQuotasIT {
         initialQuotas.setConsumerByteRate(2_000_000);
         initialQuotas.setProducerByteRate(1_000_000);
         initialQuotas.setRequestPercentage(50);
-        initialQuotas.setControllerMutationRate(10);
+        initialQuotas.setControllerMutationRate(10d);
 
         kuq.createOrUpdate(Reconciliation.DUMMY_RECONCILIATION, username, initialQuotas);
         assertThat(kuq.exists(Reconciliation.DUMMY_RECONCILIATION, username), is(true));